refactor: moved config into internal directory
